+++ extra/recorder_enumerator/test/Linq.cs
@@ -78,7 +78,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void Range([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void Range([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var range = Enumerable.Range(5, 15).GetEnumerator();
             range.MoveNext();
@@ -91,7 +91,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void WhereEnumerable([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void WhereEnumerable([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             int k = 3;
             var range = Enumerable.Range(0, 20).Where(i => i % k == 0).GetEnumerator();
@@ -105,7 +105,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void WhereArray([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void WhereArray([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             int k = 3;
             var array = Enumerable.Range(0, 20).ToArray();
@@ -120,7 +120,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void WhereList([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void WhereList([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             int k = 3;
             var list = Enumerable.Range(0, 20).ToList();
@@ -135,7 +135,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void WhereListOutOfDate([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void WhereListOutOfDate([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var list = Enumerable.Range(0, 5).ToList();
             var range = list.GetEnumerator();
@@ -156,7 +156,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void WhereSelectEnumerable([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void WhereSelectEnumerable([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             int k = 3;
             int m = 2;
@@ -171,7 +171,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void WhereSelectArray([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void WhereSelectArray([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             int k = 3;
             int m = 2;
@@ -187,7 +187,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void WhereSelectList([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void WhereSelectList([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             int k = 3;
             int m = 2;
@@ -203,7 +203,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void ObjectEnumerable([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void ObjectEnumerable([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var list = new List<StubRecordable>();
             list.Add(new StubRecordable());
@@ -217,7 +217,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void SelectEnumerable([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void SelectEnumerable([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var range = FakeRange().Select(i => i.ToString()).GetEnumerator();
             range.MoveNext();
@@ -230,7 +230,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void SelectRange([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void SelectRange([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var range = Enumerable.Range(0, 20).Select(i => i.ToString()).GetEnumerator();
             range.MoveNext();
@@ -243,7 +243,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void SelectArray([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void SelectArray([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var array = Enumerable.Range(0, 20).ToArray();
             var range = array.Select(i => i.ToString()).GetEnumerator();
@@ -257,7 +257,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void SelectList([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void SelectList([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var list = Enumerable.Range(0, 20).ToList();
             var range = list.Select(i => i.ToString()).GetEnumerator();
@@ -271,7 +271,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void SelectMany([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void SelectMany([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var range = Enumerable.Range(0, 5).SelectMany(i => Enumerable.Range(0, i)).GetEnumerator();
             range.MoveNext();
@@ -285,7 +285,7 @@ namespace RecorderEnumeratorTest
 
         [Test]
         [Ignore("This is hard to support due to Join using Lookup and Grouping.")]
-        public void JoinEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void JoinEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var outer = Enumerable.Range(1, 10);
             var inner = Enumerable.Range(1, 20);
@@ -305,7 +305,7 @@ namespace RecorderEnumeratorTest
 
         [Test]
         [Ignore("This is hard to support due to Join using Lookup and Grouping.")]
-        public void GroupJoinEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void GroupJoinEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var outer = Enumerable.Range(1, 3);
             var inner = Enumerable.Range(1, 5);
@@ -324,7 +324,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void DistinctEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void DistinctEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 40).Select(i => i % 5).Distinct().GetEnumerator();
             source.MoveNext();
@@ -335,7 +335,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void DistinctByEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void DistinctByEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 40).Select(i => i % 5).DistinctBy(x => x % 4).GetEnumerator();
             source.MoveNext();
@@ -346,7 +346,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void UnionEnumerator2Test([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void UnionEnumerator2Test([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var first = Enumerable.Range(0, 20).Select(x => x * 2);
             var second = Enumerable.Range(0, 20).Select(x => x * 3);
@@ -359,7 +359,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void UnionEnumerator3Test([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void UnionEnumerator3Test([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var first = Enumerable.Range(0, 20).Select(x => x * 2);
             var second = Enumerable.Range(0, 20).Select(x => x * 3);
@@ -373,7 +373,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void IntersectEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void IntersectEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var first = Enumerable.Range(0, 20);
             var second = Enumerable.Range(15, 20);
@@ -386,7 +386,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void IntersectByEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void IntersectByEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var first = Enumerable.Range(0, 20);
             var second = Enumerable.Range(15, 20);
@@ -399,7 +399,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void ExceptEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void ExceptEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var first = Enumerable.Range(0, 20);
             var second = Enumerable.Range(15, 20);
@@ -412,7 +412,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void ExceptByEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void ExceptByEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var first = Enumerable.Range(0, 20);
             var second = Enumerable.Range(15, 20);
@@ -425,7 +425,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void OrderByEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void OrderByEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).OrderBy(i => i % 3).GetEnumerator();
             source.MoveNext();
@@ -436,7 +436,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void OrderByDescendingEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void OrderByDescendingEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).OrderByDescending(i => i % 3).GetEnumerator();
             source.MoveNext();
@@ -447,7 +447,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void ThenByEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void ThenByEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).OrderBy(i => i % 3).ThenBy(i => i).GetEnumerator();
             source.MoveNext();
@@ -458,7 +458,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void ThenByDescendingEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void ThenByDescendingEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).OrderBy(i => i % 3).ThenByDescending(i => i).GetEnumerator();
             source.MoveNext();
@@ -469,7 +469,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void ReverseEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void ReverseEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).Reverse().GetEnumerator();
             source.MoveNext();
@@ -481,7 +481,7 @@ namespace RecorderEnumeratorTest
 
         [Test]
         [Ignore("This is hard to support due to Grouping.")]
-        public void GroupByEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void GroupByEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).GroupBy(i => i % 3).GetEnumerator();
             source.MoveNext();
@@ -492,7 +492,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void OfTypeEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void OfTypeEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var mixedSource = new List<object> { 0, 1, "two", 3, "four", 5 };
             var source = mixedSource.OfType<int>().GetEnumerator();
@@ -504,7 +504,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void CastEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void CastEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var objectSource = new List<object> { 0, 1, 2, 3, 4, 5 };
             var source = objectSource.Cast<int>().GetEnumerator();
@@ -516,7 +516,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void TakeEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void TakeEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).Take(10).GetEnumerator();
             source.MoveNext();
@@ -527,7 +527,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void SkipEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void SkipEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).Skip(10).GetEnumerator();
             source.MoveNext();
@@ -538,7 +538,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void TakeWhileEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void TakeWhileEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).TakeWhile(i => i < 10).GetEnumerator();
             source.MoveNext();
@@ -549,7 +549,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void SkipWhileEnumeratorTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void SkipWhileEnumeratorTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var source = Enumerable.Range(0, 20).SkipWhile(i => i < 10).GetEnumerator();
             source.MoveNext();
@@ -560,7 +560,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void ConcatEnumerator2Test([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void ConcatEnumerator2Test([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var first = Enumerable.Range(0, 20);
             var second = Enumerable.Range(20, 20);
@@ -573,7 +573,7 @@ namespace RecorderEnumeratorTest
         }
 
         [Test]
-        public void ConcatEnumeratorNTest([ValuesExcept(RecorderMode.Validation)] RecorderMode recorderMode)
+        public void ConcatEnumeratorNTest([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode recorderMode)
         {
             var first = Enumerable.Range(0, 20);
             var second = Enumerable.Range(20, 20);