+++ test/unit/ConverterWrite.cs
@@ -98,7 +98,7 @@ namespace DecTest
         }
 
         [Test]
-        public void ConverterStringRef([ValuesExcept(RecorderMode.Validation)] RecorderMode mode)
+        public void ConverterStringRef([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode mode)
         {
             UpdateTestParameters(new Dec.Config.UnitTestParameters { explicitConverters = new Type[] { typeof(ConvertedConverterString) } });
 
@@ -126,7 +126,7 @@ namespace DecTest
         }
 
         [Test]
-        public void ConverterReplacementWorking([ValuesExcept(RecorderMode.Validation)] RecorderMode mode)
+        public void ConverterReplacementWorking([ValuesExcept(RecorderMode.Validation, RecorderMode.Simple)] RecorderMode mode)
         {
             UpdateTestParameters(new Dec.Config.UnitTestParameters { explicitConverters = new Type[] { typeof(ConvertedConverterRecord) } });